Fixed wrong cause of death messages
The cause of death message displayed was previously always one death off (meaning that it displayed the PREVIOUS death's cause).

This was because the when retrieving the reference to the CauseOfDeathLabel the DeathScreen actually got a reference to the prefab of the CauseOfDeathLabel because the DeathScreen did not check the objects recieved for instantiation. This meant that new cause of death messages were written to the prefab and therefore, future instances of the CauseOfDeathLabel (on the next DeathScreen) had the message but not the current instance.